feat: consistent web response and error handler (#23)
* feat: consistent error handler and response

* feat: consistent error handler and response

* feat: consistent response

* feat: consistent response

* feat: consistent response

+++ product-service/internal/repository/product_repository.go
@@ -20,7 +20,7 @@ type ProductRepositoryInterface interface {
 	FindByID(ctx context.Context, id uuid.UUID) (*entity.Product, error)
 
 	// FindAll retrieves all products with optional pagination
-	FindAll(ctx context.Context, limit, offset int) ([]*entity.Product, error)
+	FindAll(ctx context.Context, limit, offset int64) ([]*entity.Product, error)
 
 	// Update updates an existing product
 	Update(ctx context.Context, product *entity.Product) (*entity.Product, error)
@@ -178,7 +178,7 @@ func (r *ProductRepositoryPostgres) FindByID(
 // FindAll finds all products with pagination.
 func (r *ProductRepositoryPostgres) FindAll(
 	ctx context.Context,
-	limit, offset int,
+	limit, offset int64,
 ) ([]*entity.Product, error) {
 	query := `
 		SELECT id, name, price, quantity, created_at, updated_at
